rename 'export radio files' & 'files util' scripts to not have underscores
