+++ fileMgt/common/filesUtil.py
