+++ fileMgt/exportRadioFiles.py
@@ -5,7 +5,7 @@ export radio files generated by sandisk sansa fuze
 
 example:
 
-    ./export_radio_files.py -s ../source/ -t ../target/
+    ./exportRadioFiles.py -s ../source/ -t ../target/
 
 notes:
 
@@ -26,7 +26,7 @@ import os
 from time import mktime
 import sys
 sys.path.append('./common')
-import files_util
+import filesUtil
 import argparse
 
 # command line args
@@ -44,7 +44,7 @@ if source_dir == target_dir:
     print('Unable to move files. Source and target directories are the same')
     sys.exit()
 
-if not files_util.dirs_exist([source_dir, target_dir]):
+if not filesUtil.dirs_exist([source_dir, target_dir]):
     print('Unable to move files. Source and/or target directories do not exist.')
     sys.exit()
 
@@ -72,4 +72,4 @@ else:
 
 # move new files to target directory
 
-files_util.move_files(source_dir, target_dir, FILE_EXT, "radio")
+filesUtil.move_files(source_dir, target_dir, FILE_EXT, "radio")
